feat: send ending message to Africa's Talking
Africa's Talking needs to know when they should end the user's session.
Unfournately, Botman does not include support for this kind of interaction.
This adds a new type of message, EndingMesage, which extends and behaves
the same way Botman's OutgoingMessage does, but it can be used by the driver
to know that's the ending message of the session.

Also, in order to add this custom behavior, we start implementing more methods
from the HttpDriver interface in the AfricasTalkingDriver, withoud need of the WebDriver
intervention.